Rewrote a most of DeadmanConfig to improve API design
Config entries are now defined by an instance of either ConfigEntry,
ListConfigEntry, or MapConfigEntry of rather than by enum constants.
Each of these classes handle the loading of different EntryValue types;
Single, List, and Map entry values.  Each of these config entry classes
extend BaseConfigEntry which defines the public methods to get the value
class type, the path to the entry, the actual value, and whether that
value is the default.