Cleaned up DeadmanUtils and added isNumeric method
The getLocationFromKey and formatLocationKey DeadmanUtils methods were
removed since that functionality is now apart of the WorldCoord class in
its parse and toStirng/toCompactString methods.  An isNumeric method was
added that simply checks if the String is not null or empty then uses
Apache Commons StringUtils.isNumeric because that method before commons v3
returns true for empty strings for some stupid reason.