Renames PlayerID to PlayerId and improves it so the username is updated with the most recent
Previously, this class was fully immutable which meant the username would never change unless
a new instance was made with an updated username.  Now, calling the getUsername method
will check if the name returned by getOfflinePlayer().getName() is not null and is different from the
current username and update it accordingly.  This way, a plugin can continue to serialize and load
a PlayerId instance without worrying about stale usernames.